requirements.production.txt
# Dependências essenciais para produção - Youth Space Backend
# Versões fixas para estabilidade em produção

# Framework Flask
Flask==2.3.2
Werkzeug==3.1.3

# Banco de dados
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
SQLAlchemy==2.0.43
PyMySQL==1.1.0
mysql-connector-python==9.3.0

# Autenticação e segurança
Flask-JWT-Extended==4.6.0
PyJWT==2.10.1
bcrypt==4.0.1

# CORS
flask-cors==6.0.1

# Configuração
python-dotenv==1.0.0

# Servidor WSGI para produção
gunicorn==21.2.0

# Utilitários essenciais
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
blinker==1.9.0

# Migrations
alembic==1.16.5
Mako==1.3.10

# Comunicação externa
requests==2.32.3
urllib3==2.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10

# Email e SMS (opcionais)
twilio==9.8.0

# Processamento de dados (se necessário)
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.17.0

# Sistema
colorama==0.4.6
typing_extensions==4.13.2

# Monitoramento e logs (opcionais)
# sentry-sdk[flask]==1.32.0  # Descomente se usar Sentry
# newrelic  # Descomente se usar New Relic
